requirements.txt
agate==1.7.0
attrs==25.3.0
babel==2.17.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dbt-core==1.5.10
dbt-extractor==0.4.1
dbt-greenplum==1.5.0
dbt-postgres==1.5.10
future==1.0.0
greenlet==3.2.2
hologram==0.0.16
idna==3.10
isodate==0.6.1
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
leather==0.4.0
Logbook==1.5.3
MarkupSafe==3.0.2
mashumaro==3.6
minimal-snowplow-tracker==0.0.2
msgpack==1.1.0
networkx==2.8.8
numpy==2.2.6
packaging==25.0
pandas==2.2.3
parsedatetime==2.4
pathspec==0.11.2
protobuf==6.31.0
psycopg2-binary==2.9.10
pycparser==2.22
python-dateutil==2.8.2
python-dotenv==1.1.0
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
setuptools==80.9.0
six==1.17.0
SQLAlchemy==2.0.41
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==2.3.8
